Added tests for /ListSimulations command line option.
