+++ Tests/UnitTests/UnitTests.csproj
@@ -146,7 +146,6 @@
     <Compile Include="APSIMShared\MockJob.cs" />
     <Compile Include="CommandLineArgsTests.cs" />
     <Compile Include="Core\ApsimFile\EditFileTests.cs" />
-    <Compile Include="Core\CommandLineOptionsTests.cs" />
     <Compile Include="Core\Run\GenerateApsimXFilesTests.cs" />
     <Compile Include="Core\Run\RunnerTests.cs" />
     <Compile Include="Core\Run\SimulationTests.cs" />
@@ -277,6 +276,7 @@
     <EmbeddedResource Include="Report\ReportOnEvents.apsimx" />
     <EmbeddedResource Include="Resources\WheatModel.apsimx" />
     <EmbeddedResource Include="Stock\AddAnimalCohort.json" />
+    <EmbeddedResource Include="Resources\SimpleExperiment.apsimx" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\APSIM.Shared\APSIM.Shared.csproj">
