+++ Tests/UnitTests/Utilities.cs
@@ -5,6 +5,7 @@ namespace UnitTests
     using APSIM.Shared.Utilities;
     using Models;
     using Models.Core;
+    using Models.Core.ApsimFile;
     using Models.GrazPlan;
     using Models.Storage;
     using System;
@@ -192,6 +193,21 @@ namespace UnitTests
             Apsim.ParentAllChildren(sims);
             sims.Write(sims.FileName);
             return sims;
+        }
+
+        public static Simulations GetSimpleExperiment()
+        {
+            Simulations result = ReadFromResource<Simulations>("UnitTests.Resources.SimpleExperiment.apsimx", out List<Exception> errors);
+            if (errors != null && errors.Count > 0)
+                throw errors[0];
+
+            return result;
+        }
+
+        public static T ReadFromResource<T>(string resourceName, out List<Exception> creationExceptions) where T : IModel
+        {
+            string json = ReflectionUtilities.GetResourceAsString(resourceName);
+            return FileFormat.ReadFromString<T>(json, out creationExceptions);
         }
 
         /// <summary>
